dir for queryPriceHistory
