+++ backend/functions/queryPriceHistory/requirements.txt
@@ -1,0 +1,1 @@
+appwrite==4.1.0
